+++ packages/api/src/handlers/api-gateway/feed/actions/__tests__/create-feed-action.test.ts
@@ -1,12 +1,8 @@
 import type { AppContext } from '@/app/chase-light-app'
 import { StubTokenParser } from '@/features/auth/services/stub-token-parser'
 import { swapTokenParserForTest } from '@/features/auth/services/token-parser'
-import type {
-  SignupResult,
-  SignupViaProvider,
-  User,
-} from '@/features/user/domain/user'
-import { userApp } from '@/features/user/functions'
+import type { CreateFeedRequest, Feed } from '@/features/feed/domain/feed'
+import { feedApp } from '@/handlers/api-gateway/feed'
 import type { OpenAPIHono } from '@hono/zod-openapi'
 import { UserFactory } from 'prisma/seeds/user-factory'
 import { beforeEach, describe, expect, test } from 'vitest'
@@ -14,28 +10,33 @@ import { beforeEach, describe, expect, test } from 'vitest'
 let app: OpenAPIHono<AppContext>
 
 beforeEach(() => {
-  app = userApp.getApp()
+  app = feedApp.getApp()
 })
 
-describe('SignupVieProviderAction', () => {
-  test.sequential('新規登録テスト', async () => {
+describe('CreateFeedAction', () => {
+  test('フィードを作成できること', async () => {
     const stubTokenParser = new StubTokenParser()
-    const user = (await UserFactory.build()) as User
+    const user = await UserFactory.create()
 
     stubTokenParser.setUser(user)
     swapTokenParserForTest(stubTokenParser)
 
-    const result = await app.request('/users/signup-via-provider', {
+    const request: CreateFeedRequest = {
+      name: 'test feed',
+      url: 'https://github.com/abc/def',
+      cycle: 1, // CYCLE_VALUE_MAP.DAILY
+    }
+
+    const result = await app.request('/feeds', {
       method: 'POST',
       headers: {
         'Content-Type': 'application/json',
       },
-      body: JSON.stringify({
-        accessToken: '',
-        idToken: '',
-      } satisfies SignupViaProvider),
+      body: JSON.stringify(request),
     })
     expect(result.status).toBe(200)
-    expect(((await result.json()) as SignupResult).success).toBe(true)
+    expect(((await result.json()) as { feed: Feed }).feed.name).toBe(
+      request.name,
+    )
   })
 })
