+++ packages/api/src/handlers/api-gateway/user/actions/__tests__/fetch-user-action.test.ts
@@ -2,7 +2,7 @@ import type { AppContext } from '@/app/chase-light-app'
 import { StubTokenParser } from '@/features/auth/services/stub-token-parser'
 import { swapTokenParserForTest } from '@/features/auth/services/token-parser'
 import type { User } from '@/features/user/domain/user'
-import { userApp } from '@/features/user/functions'
+import { userApp } from '@/handlers/api-gateway/user'
 import type { OpenAPIHono } from '@hono/zod-openapi'
 import { UserFactory } from 'prisma/seeds/user-factory'
 import { beforeEach, describe, expect, test } from 'vitest'
