+++ packages/api/src/handlers/api-gateway/user/actions/__tests__/signup-via-provider-action.test.ts
@@ -1,42 +1,41 @@
 import type { AppContext } from '@/app/chase-light-app'
 import { StubTokenParser } from '@/features/auth/services/stub-token-parser'
 import { swapTokenParserForTest } from '@/features/auth/services/token-parser'
-import type { feedSearchResult } from '@/features/feed/domain/feed'
-import { feedApp } from '@/features/feed/functions'
+import type {
+  SignupResult,
+  SignupViaProvider,
+  User,
+} from '@/features/user/domain/user'
+import { userApp } from '@/handlers/api-gateway/user'
 import type { OpenAPIHono } from '@hono/zod-openapi'
-import { FeedFactory } from 'prisma/seeds/feed-factory'
 import { UserFactory } from 'prisma/seeds/user-factory'
 import { beforeEach, describe, expect, test } from 'vitest'
 
 let app: OpenAPIHono<AppContext>
 
 beforeEach(() => {
-  app = feedApp.getApp()
+  app = userApp.getApp()
 })
 
-describe('ListFeedAction', () => {
-  test('フィード一覧を取得できること', async () => {
+describe('SignupVieProviderAction', () => {
+  test.sequential('新規登録テスト', async () => {
     const stubTokenParser = new StubTokenParser()
-    const user = await UserFactory.create()
-    await FeedFactory.createList(5, {
-      user: {
-        connect: {
-          id: user.id,
-        },
-      },
-    })
+    const user = (await UserFactory.build()) as User
 
     stubTokenParser.setUser(user)
     swapTokenParserForTest(stubTokenParser)
 
-    const result = await app.request('/feeds', {
-      method: 'GET',
+    const result = await app.request('/users/signup-via-provider', {
+      method: 'POST',
       headers: {
         'Content-Type': 'application/json',
       },
+      body: JSON.stringify({
+        accessToken: '',
+        idToken: '',
+      } satisfies SignupViaProvider),
     })
-    const resultJson = (await result.json()) as feedSearchResult
-    expect(resultJson.result.length).toBe(5)
     expect(result.status).toBe(200)
+    expect(((await result.json()) as SignupResult).success).toBe(true)
   })
 })
