+++ packages/api/src/handlers/api-gateway/user/actions/fetch-user-action.ts
