+++ packages/api/src/handlers/api-gateway/user/actions/signup-via-provider-action.ts
