+++ packages/api/src/handlers/api-gateway/user/index.ts
@@ -1,6 +1,6 @@
 import { ChaseLightApp } from '@/app/chase-light-app'
-import { FetchSelfAction } from '@/features/user/functions/actions/fetch-user-action'
-import { SignupVieProviderAction } from '@/features/user/functions/actions/signup-via-provider-action'
+import { FetchSelfAction } from '@/handlers/api-gateway/user/actions/fetch-user-action'
+import { SignupVieProviderAction } from '@/handlers/api-gateway/user/actions/signup-via-provider-action'
 import { handlerPath } from '@/lib/hono/handler-resolver'
 import { currentDirPath } from '@/lib/utils/path-utils'
 import {
